Three part version for olde PSGet and tell others to get newer one

+++ publish.ps1
@@ -7,6 +7,15 @@ param(
 
 $ErrorActionPreference = "STOP"
 
+Import-Module PowerShellGet -ErrorAction SilentlyContinue
+$powershellGet = Get-Module PowerShellGet
+if(-not $powershellGet) {
+    Write-Warning "Could not find the PowerShellGet module. 'Install-Module PowerShellGet' first."
+} elseif($powershellGet.Version -lt "1.6.0") {
+    Write-Warning "PowerShellGet 1.6.0 and greater is required. 'Install-Module PowerShellGet -Force' first."
+    exit 1
+}
+
 if(-not $Locally -and -not $PowerShellGallery) {
     Write-Warning "You must specify either -Locally or -PowerShellGallery"
     exit 1
@@ -17,7 +26,9 @@ if($Locally) {
     $name = "LocalBigIp"
     $path = "$PSScriptRoot\localgallery"
 
-    Remove-Item $path -Recurse -Force | Out-Null
+    if((Test-Path $path)) {
+        Remove-Item $path -Recurse -Force | Out-Null
+    }
     New-Item -Type Directory $path -Force | Out-Null
     Register-PSRepository -Name $name -SourceLocation $path -PublishLocation $path -InstallationPolicy Trusted
